Add files via upload
The ride_test_two.py script is the main one which is based on sequential driving for coordinate starting at 1 and ending with 8. I've been trying different values for left and right turns, but there are no optimum ones for now. I'm using timers for coordinates. In the main script, there are variables named coordinate one and eight. The self.red_counter pauses the timer to execute the next function when the red light is on. Team members are currently working on lane detection, roundabout, and collision avoidance via lidar. Note the subprocess function to change LED strip creates 500 ms delay, which can result in the next command be skipped. This an error to be tackled.

The stop sign and road cones get detected by the nodes separately. Through ROS2 topics, the depths of the road cone and stop signs can be easily communicated from the node in one script to the other node in other script.